Move ratelimit out of gateway package (#2384)
